Break off error to own module
